requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements_lock.txt requirements.in
#
attrs==23.1.0
    # via fiona
audioread==3.0.1
    # via librosa
branca==0.6.0
    # via folium
certifi==2023.7.22
    # via
    #   fiona
    #   pyproj
    #   requests
cffi==1.16.0
    # via soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r requirements.in
    #   click-plugins
    #   cligj
    #   fiona
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via
    #   librosa
    #   moviepy
filelock==3.16.1
    # via torch
fiona==1.9.5
    # via geopandas
folium==0.14.0
    # via -r requirements.in
fonttools==4.55.3
    # via matplotlib
fsspec==2024.12.0
    # via torch
geopandas==0.14.0
    # via -r requirements.in
idna==3.4
    # via requests
imageio==2.34.0
    # via moviepy
imageio-ffmpeg==0.4.9
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   branca
    #   folium
    #   torch
joblib==1.4.0
    # via
    #   librosa
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.10.1
    # via -r requirements.in
llvmlite==0.42.0
    # via numba
markupsafe==2.1.3
    # via jinja2
matplotlib==3.10.0
    # via
    #   seaborn
    #   ultralytics
moviepy==0.2.3.1
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
msgpack==1.0.8
    # via librosa
networkx==3.4.2
    # via torch
numba==0.59.1
    # via librosa
numpy==1.26.1
    # via
    #   -r requirements.in
    #   contourpy
    #   folium
    #   imageio
    #   librosa
    #   matplotlib
    #   moviepy
    #   numba
    #   opencv-python
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   shapely
    #   soxr
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.10.0.84
    # via
    #   -r requirements.in
    #   ultralytics
packaging==23.2
    # via
    #   geopandas
    #   lazy-loader
    #   matplotlib
    #   pooch
    #   pytest
pandas==2.1.2
    # via
    #   -r requirements.in
    #   geopandas
    #   seaborn
    #   ultralytics
piexif==1.1.3
    # via -r requirements.in
pillow==10.3.0
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   torchvision
    #   ultralytics
platformdirs==4.2.0
    # via pooch
pluggy==1.5.0
    # via pytest
pooch==1.8.1
    # via librosa
protobuf==5.27.1
    # via -r requirements.in
psutil==6.1.1
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pycparser==2.22
    # via cffi
pyparsing==3.2.1
    # via matplotlib
pyproj==3.6.1
    # via geopandas
pytest==8.2.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytsk3==20231007
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.2
    # via ultralytics
requests==2.31.0
    # via
    #   -r requirements.in
    #   folium
    #   pooch
    #   ultralytics
scikit-learn==1.4.2
    # via librosa
scipy==1.13.0
    # via
    #   librosa
    #   scikit-learn
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
shapely==2.0.2
    # via geopandas
six==1.16.0
    # via
    #   fiona
    #   python-dateutil
soundfile==0.12.1
    # via librosa
soxr==0.3.7
    # via librosa
sympy==1.13.1
    # via torch
threadpoolctl==3.4.0
    # via scikit-learn
torch==2.5.1
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.20.1
    # via ultralytics
tqdm==4.66.2
    # via
    #   moviepy
    #   ultralytics
typing-extensions==4.11.0
    # via
    #   librosa
    #   torch
tzdata==2023.3
    # via pandas
ultralytics==8.3.58
    # via -r requirements.in
ultralytics-thop==2.0.13
    # via ultralytics
urllib3==2.0.7
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
